toybench/requirements.txt
google-generativeai
python-dotenv
selenium
requests